+++ Source/Krypton Components/Krypton.Toolkit/Controls Toolkit/KryptonManager.cs
@@ -1586,6 +1586,18 @@ public sealed class KryptonManager : Component
                 return PaletteSparkleDeuteranopia;
             case PaletteMode.SparkleProtanopia:
                 return PaletteSparkleProtanopia;
+            case PaletteMode.Office2007LimeGreen:
+                return PaletteOffice2007LimeGreen;
+            case PaletteMode.Office2007LimeGreenDark:
+                return PaletteOffice2007LimeGreenDark;
+            case PaletteMode.Office2010LimeGreen:
+                return PaletteOffice2010LimeGreen;
+            case PaletteMode.Office2010LimeGreenDark:
+                return PaletteOffice2010LimeGreenDark;
+            case PaletteMode.Microsoft365LimeGreen:
+                return PaletteMicrosoft365LimeGreen;
+            case PaletteMode.Microsoft365LimeGreenDark:
+                return PaletteMicrosoft365LimeGreenDark;
 
             case PaletteMode.Custom:
             case PaletteMode.Global:
@@ -1951,6 +1963,34 @@ public sealed class KryptonManager : Component
 
     /// <summary>Gets the Sparkle protanopia accessibility palette.</summary>
     public static PaletteSparkleProtanopia PaletteSparkleProtanopia => _paletteSparkleProtanopia ??= new PaletteSparkleProtanopia();
+    /// Gets the single instance of the light Lime Green variant Office 2007 palette.
+    /// </summary>
+    public static PaletteOffice2007LimeGreen PaletteOffice2007LimeGreen => _paletteOffice2007LimeGreen ??= new PaletteOffice2007LimeGreen();
+
+    /// <summary>
+    /// Gets the single instance of the dark Lime Green variant Office 2007 palette.
+    /// </summary>
+    public static PaletteOffice2007LimeGreenDark PaletteOffice2007LimeGreenDark => _paletteOffice2007LimeGreenDark ??= new PaletteOffice2007LimeGreenDark();
+
+    /// <summary>
+    /// Gets the single instance of the light Lime Green variant Office 2010 palette.
+    /// </summary>
+    public static PaletteOffice2010LimeGreen PaletteOffice2010LimeGreen => _paletteOffice2010LimeGreen ??= new PaletteOffice2010LimeGreen();
+
+    /// <summary>
+    /// Gets the single instance of the dark Lime Green variant Office 2010 palette.
+    /// </summary>
+    public static PaletteOffice2010LimeGreenDark PaletteOffice2010LimeGreenDark => _paletteOffice2010LimeGreenDark ??= new PaletteOffice2010LimeGreenDark();
+
+    /// <summary>
+    /// Gets the single instance of the light Lime Green variant Microsoft 365 palette.
+    /// </summary>
+    public static PaletteMicrosoft365LimeGreen PaletteMicrosoft365LimeGreen => _paletteMicrosoft365LimeGreen ??= new PaletteMicrosoft365LimeGreen();
+
+    /// <summary>
+    /// Gets the single instance of the dark Lime Green variant Microsoft 365 palette.
+    /// </summary>
+    public static PaletteMicrosoft365LimeGreenDark PaletteMicrosoft365LimeGreenDark => _paletteMicrosoft365LimeGreenDark ??= new PaletteMicrosoft365LimeGreenDark();
 
     private static PaletteMaterialLight? _paletteMaterialLight;
     private static PaletteMaterialDark? _paletteMaterialDark;
@@ -1977,6 +2017,13 @@ public sealed class KryptonManager : Component
     private static PaletteSparkleHighContrast? _paletteSparkleHighContrast;
     private static PaletteSparkleDeuteranopia? _paletteSparkleDeuteranopia;
     private static PaletteSparkleProtanopia? _paletteSparkleProtanopia;
+
+    private static PaletteOffice2007LimeGreen? _paletteOffice2007LimeGreen;
+    private static PaletteOffice2007LimeGreenDark? _paletteOffice2007LimeGreenDark;
+    private static PaletteOffice2010LimeGreen? _paletteOffice2010LimeGreen;
+    private static PaletteOffice2010LimeGreenDark? _paletteOffice2010LimeGreenDark;
+    private static PaletteMicrosoft365LimeGreen? _paletteMicrosoft365LimeGreen;
+    private static PaletteMicrosoft365LimeGreenDark? _paletteMicrosoft365LimeGreenDark;
 
     //public static PaletteBase CustomPaletteBase => _customPalette ??= new PaletteBase ();
 
@@ -2240,6 +2287,8 @@ public sealed class KryptonManager : Component
             case PaletteMode.Office2007White:
             case PaletteMode.Office2007Black:
             case PaletteMode.Office2007BlackDarkMode:
+            case PaletteMode.Office2007LimeGreen:
+            case PaletteMode.Office2007LimeGreenDark:
             case PaletteMode.VisualStudio2010Render2007:
             case PaletteMode.Office2007HighContrast:
             case PaletteMode.Office2007Deuteranopia:
@@ -2267,6 +2316,8 @@ public sealed class KryptonManager : Component
             case PaletteMode.SparkleHighContrast:
             case PaletteMode.SparkleDeuteranopia:
             case PaletteMode.SparkleProtanopia:
+            case PaletteMode.Office2010LimeGreen:
+            case PaletteMode.Office2010LimeGreenDark:
             case PaletteMode.VisualStudio2010Render2010:
             case PaletteMode.Office2010HighContrast:
             case PaletteMode.Office2010Deuteranopia:
@@ -2292,6 +2343,8 @@ public sealed class KryptonManager : Component
             case PaletteMode.Microsoft365SilverDarkMode:
             case PaletteMode.Microsoft365SilverLightMode:
             case PaletteMode.Microsoft365White:
+            case PaletteMode.Microsoft365LimeGreen:
+            case PaletteMode.Microsoft365LimeGreenDark:
             case PaletteMode.VisualStudio2010Render365:
             case PaletteMode.HighContrast:
             case PaletteMode.Deuteranopia: