+++ Source/Krypton Components/Krypton.Toolkit/Controls Toolkit/KryptonThemeComboBox.cs
@@ -90,8 +90,46 @@ public class KryptonThemeComboBox : KryptonComboBox, IKryptonThemeSelectorBase
     {
         // React to theme changes from outside this control.
         KryptonManager.GlobalPaletteChanged += KryptonManagerGlobalPaletteChanged;
+        ThemeManager.RegisteredThemesChanged += ThemeManagerRegisteredThemesChanged;
         base.OnHandleCreated(e);
     }
+
+    /// <inheritdoc />
+    protected override void OnHandleDestroyed(EventArgs e)
+    {
+        KryptonManager.GlobalPaletteChanged -= KryptonManagerGlobalPaletteChanged;
+        ThemeManager.RegisteredThemesChanged -= ThemeManagerRegisteredThemesChanged;
+        base.OnHandleDestroyed(e);
+    }
+
+    private void ThemeManagerRegisteredThemesChanged(object? sender, EventArgs e)
+    {
+        if (IsDisposed)
+        {
+            return;
+        }
+
+        int previous = SelectedIndex;
+        Items.Clear();
+        Items.AddRange(CommonHelperThemeSelectors.GetThemesArray());
+        if (previous >= 0 && previous < Items.Count)
+        {
+            _isExternalUpdate = true;
+            try
+            {
+                SelectedIndex = previous;
+            }
+            finally
+            {
+                _isExternalUpdate = false;
+            }
+        }
+        else
+        {
+            SelectedIndex = CommonHelperThemeSelectors.GetPaletteIndex(Items, KryptonManager.CurrentGlobalPaletteMode);
+        }
+    }
+
     /// <summary>
     /// This method will run when the KryptonManager.GlobalPaletteChanged event is fired.<br/>
     /// It will synchronize the SelectedIndex with the newly assigned Global Palette.